reorganizacion de clase board y game
